+++ src/georef_ar_py/__init__.py
@@ -1,0 +1,5 @@
+__version__ = '0.0.2'
+import logging
+
+if __name__ == '__main__':
+    logging.basicConfig(format='%(asctime)s %(levelname)8s %(message)s')